refactor: make useFriendRequest hook
