+++ front/src/components/NavigationBar/DefaultNavBtnList/FriendRequestContainer/FriendRequestList/FriendRequest/FriendRequest.js
@@ -1,12 +1,7 @@
-import React, { useContext } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
-import { useMutation } from '@apollo/react-hooks';
 import Message from 'components/globalComponents/Message/Message';
-import {
-  ACCEPT_FRIEND_REQUEST,
-  DELETE_FRIEND_REQUEST,
-} from 'queries/beforeFriend';
-import GlobalContext from 'global.context';
+import useFriendRequest from 'hooks/FriendsSection/useFriendRequest';
 
 FriendRequest.propTypes = {
   id: PropTypes.number.isRequired,
@@ -15,29 +10,8 @@ FriendRequest.propTypes = {
 };
 
 export default function FriendRequest({ id, nickname, remove }) {
-  const { onlineSocket } = useContext(GlobalContext);
-  const [deleteFriendRequest] = useMutation(DELETE_FRIEND_REQUEST, {
-    onCompleted() {
-      remove();
-    },
-  });
+  const [acceptRequest, declineRequest] = useFriendRequest({ id, remove });
 
-  const [acceptFriendRequest] = useMutation(ACCEPT_FRIEND_REQUEST, {
-    onCompleted({ acceptFriendRequest: friend }) {
-      onlineSocket.emit('acceptFriendRequest', friend);
-      remove();
-    },
-  });
-
-  const acceptRequest = async () => {
-    await acceptFriendRequest({ variables: { id } });
-  };
-
-  const declineRequest = async () => {
-    await deleteFriendRequest({
-      variables: { id },
-    });
-  };
   return (
     <Message
       content={`${nickname}님이 친구 요청을 하였습니다.`}
